made clean_emulator.sh more robust, can now work while jobs are running

+++ clean_emulator.sh
@@ -1,9 +1,24 @@
 #!/bin/bash
 
-# NOTE can only run this while emulator sample creation is not taking place!!!
+# NOTE this also works while jobs are running now!
+
+now_time=$(date +'%s')
+
+# let at least 2 hours be elapsed since creation so we know
+# for sure nobody is working on this right now
+min_time_delta=$(( 2 * 60 * 60 ))
 
 root='/scratch/gpfs/lthiele/nuvoid_production'
 for d in "$root"/cosmo_varied_*/emulator/*; do
+  # check for stale/work in progress
+  if [ ! -f $d/hod.info ]; then continue; fi
+
+  create_time=$(date +'%s' -r $d/hod.info)
+
+  # if the directory is relatively recent, do not touch
+  # (someone may be working on it)
+  if [ $(( now_time-create_time )) -lt $min_time_delta ]; then continue; fi
+
   rm -f "$d"/lightcone_*.txt &
   rm -rf "$d"/logs_* &
   rm -rf "$d/figs" &
